backend/requirements.txt
ultralytics==8.2.0
Pillow==10.0.0
opencv-python==4.10.0.84
